requirements.txt
aiogram~=3.20.0
dotenv~=0.9.9
python-dotenv~=1.1.0
tabula-py~=2.10.0
pandas~=2.2.3